# [Performance](func) opt the print_id func