## Proposed changes

```
Load Average: 52.48, 35.97, 38.43
------------------------------------------------------
Benchmark            Time             CPU   Iterations
------------------------------------------------------
old 3390427270 ns   3390354519 ns            1
new    335514305 ns    335513720 ns            2
```

<!--Describe your changes.-->